requirements.txt
mcp
proxmoxer
requests
aiohttp
python-dotenv
configparser
apscheduler
gotify